Review: validationMode refactor, bestEfortJsonSchemaEncoder in json-utils
