+++ docs/MigrationGuide.md
@@ -15,6 +15,9 @@ To see the biggest differences please consult the [changelog](Changelog.md).
 * [#3610](https://github.com/TouK/nussknacker/pull/3610) Removed deprecated code. For details see changes in pull request.
 * [#3576](https://github.com/TouK/nussknacker/pull/3576) `/processes` endpoint without query parameters returns all scenarios - the previous behaviour was to return only unarchived ones.
   To fetch only unarchived scenarios `isArchived=false` query parameter has to be passed.
+* [#3607](https://github.com/TouK/nussknacker/pull/3607) Request-response jsonSchema based encoder:
+  * ValidationMode moved to package `pl.touk.nussknacker.engine.api.validation` in `nussknacker-components-api`
+  * BestEffortJsonSchemaEncoder moved to package `pl.touk.nussknacker.engine.json.encode` in `nussknacker-json-utils`
 
 ## In version 1.6.0
 * [#3440](https://github.com/TouK/nussknacker/pull/3440) Feature: allow to define fragment's outputs
