+++ engine/lite/request-response/runtime/src/main/scala/pl/touk/nussknacker/engine/requestresponse/DefaultResponseEncoder.scala
@@ -1,19 +1,14 @@
 package pl.touk.nussknacker.engine.requestresponse
 
 import io.circe.Json
-import org.everit.json.schema.Schema
 import pl.touk.nussknacker.engine.requestresponse.api.ResponseEncoder
-import pl.touk.nussknacker.engine.schemedkafka.encode.{BestEffortJsonSchemaEncoder, ValidationMode}
+import pl.touk.nussknacker.engine.util.json.BestEffortJsonEncoder
 
 object DefaultResponseEncoder extends ResponseEncoder[Any] {
 
-  private val bestEffortEncoder = new BestEffortJsonSchemaEncoder(ValidationMode.lax)
+  private val bestEffortEncoder = BestEffortJsonEncoder(failOnUnkown = true, getClass.getClassLoader)
 
-  override def toJsonResponse(input: Any, result: List[Any], schema: Option[Schema]): Json = {
-    result
-      .map(bestEffortEncoder.encodeOrError(_, schema.get))
-      .headOption
-      .getOrElse(throw new IllegalArgumentException(s"Process did not return any result"))
-  }
+  override def toJsonResponse(input: Any, result: List[Any]): Json = bestEffortEncoder.encode(result)
 
 }
+
